+++ src/samTextViewer/Main.java
@@ -1,21 +1,17 @@
 package samTextViewer;
 
 import java.io.BufferedReader;
-import java.io.BufferedWriter;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileReader;
-import java.io.FileWriter;
 import java.io.IOException;
 import java.io.InputStreamReader;
 import java.net.MalformedURLException;
 import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.Arrays;
-import java.util.LinkedHashSet;
 import java.util.List;
 import java.util.ListIterator;
-import java.util.Set;
 
 import org.apache.commons.lang3.StringUtils;
 
@@ -40,7 +36,6 @@ import jline.console.ConsoleReader;
 import jline.console.completer.StringsCompleter;
 import jline.console.history.History;
 import jline.console.history.History.Entry;
-import jline.console.history.MemoryHistory;
 import net.sourceforge.argparse4j.inf.Namespace;
 import tracks.IntervalFeature;
 import tracks.Track;
@@ -56,9 +51,6 @@ public class Main {
 	
 	public static void main(String[] args) throws IOException, InvalidGenomicCoordsException, InvalidCommandLineException, InvalidRecordException, BamIndexNotFoundException, ClassNotFoundException, SQLException, DocumentException, UnindexableFastaFileException, InvalidColourException, InvalidConfigException {
 
-		final String CMD_HISTORY_FILE= System.getProperty("user.home") + File.separator + ".asciigenome_history";  
-		final String MARKER_FOR_HISTORY_FILE= "## file ##";
-		final String MARKER_FOR_HISTORY_CMD= "## cmd ##";
 		/* Start parsing arguments * 
 		 * *** If you change something here change also in console input ***/
 		Namespace opts= ArgParse.argParse(args);
@@ -74,6 +66,8 @@ public class Main {
 		// Get configuration. Note that we don't need to assign this to a variable. 
 		new Config(config);
 		new Xterm256();
+
+		ASCIIGenomeHistory asciiGenomeHistory= new ASCIIGenomeHistory();
 		
 		// Init console right at start so if something goes wrong the user's terminal is reset to 
 		// initial defaults with the shutdown hook. This could be achieved in cleaner way probably.
@@ -92,7 +86,6 @@ public class Main {
 		/* Initialize trackSet */
 		/* ------------------- */
 		// This part only prepares a dummy GenomicCoords object to initialize the start position:
-		// ----------------------------
 		
 		if(region == null || region.isEmpty()){
 			region= initRegion(inputFileList, fasta, null, debug);			
@@ -110,12 +103,13 @@ public class Main {
 		// Genomic positions start here:
 		final GenomicCoordsHistory gch= new GenomicCoordsHistory();
 		GenomicCoords start= new GenomicCoords(initGc.toStringRegion(), terminalWidth, initGc.getSamSeqDict(), initGc.getFastaFile());
-		gch.readHistory(new File(CMD_HISTORY_FILE), start);
+		
+		gch.readHistory(asciiGenomeHistory.getFileName(), start);
 		gch.add(start);
 
 		final TrackSet trackSet= new TrackSet(inputFileList, gch.current());
-		addHistoryFiles(trackSet, CMD_HISTORY_FILE, MARKER_FOR_HISTORY_FILE, debug);
-		
+		trackSet.addHistoryFiles(asciiGenomeHistory.getFiles());
+
 		setDefaultTrackHeights(console.getTerminal().getHeight(), trackSet.getTrackList());
 		
 		final TrackProcessor proc= new TrackProcessor(trackSet, gch);
@@ -173,9 +167,11 @@ public class Main {
 
 		/* Set up done, start processing */
 		/* ============================= */
-		History cmdHistory= initCmdHistory(CMD_HISTORY_FILE, MARKER_FOR_HISTORY_CMD, debug);
-		console.setHistory(cmdHistory);
-		writeHistory(console.getHistory(), trackSet.getOpenedFiles(), CMD_HISTORY_FILE, MARKER_FOR_HISTORY_FILE, MARKER_FOR_HISTORY_CMD, gch);
+		console.setHistory(asciiGenomeHistory.getCommandHistory());
+		writeYamlHistory(asciiGenomeHistory, 
+				         console.getHistory(), 
+				         trackSet, 
+				         gch);
 		
 		while(true){  
 			// keep going until quit or if no interactive input set
@@ -194,7 +190,7 @@ public class Main {
 
 				cmdConcatInput= console.readLine().trim();
 				if (cmdConcatInput.isEmpty()) {
-					// Empty inout: User only issued <ENTER> 
+					// Empty input: User only issued <ENTER> 
 					if( interactiveInput.getInteractiveInputExitCode().equals(ExitCode.CLEAN)){
 						// User only issued <ENTER>: Repeat previous command if the exit code was not an error.
 						cmdConcatInput= currentCmdConcatInput;					
@@ -268,37 +264,6 @@ public class Main {
 		}
 		// It appears everything failed to initialise...
 		return "";
-//		List<String>initGenomeList= new ArrayList<String>(inputFileList);
-		
-		/* Prepare genomic coordinates to fetch. This should probably be a function in itself */
-		// Create a dummy gc object just to get the sequence dict.
-//		GenomicCoords gc= new GenomicCoords(Utils.getTerminalWidth());
-//		
-//		if(genome != null && ! genome.trim().isEmpty()){
-//			initGenomeList.add(genome);
-//		}
-//		gc.setGenome(initGenomeList, false);
-//		
-//		SAMSequenceDictionary samSeqDict = gc.getSamSeqDict();
-
-//		if(samSeqDict != null && ! samSeqDict.isEmpty()){
-//			region= samSeqDict.getSequence(0).getSequenceName();
-//			System.err.println("");
-//		} else {
-//			for(String x : initGenomeList){
-//				try {
-//					region= Utils.initRegionFromFile(x);
-//					System.err.println("Done from: " + x);
-//					break;
-//				} catch(Exception e){
-//					System.err.println("\nCould not initilize from file " + x);
-//					if(debug > 0){
-//						e.printStackTrace();
-//					}
-//				}
-//			}
-//		}
-//		return region;
 	}
 	
 	/** If exec is a file, parse it to return a string suitable for execution.  
@@ -340,65 +305,6 @@ public class Main {
 		
 	}
 
-	/**Read the asciigenome history file and put it a list as current history. Or
-	 * return empty history file does not exist or can't be read. 
-	 * */
-	private static History initCmdHistory(String cmdHistoryFile, String MARKER_FOR_HISTORY_CMD, int debug){
-		History cmdHistory= new MemoryHistory();
-		try{
-			BufferedReader br= new BufferedReader(new FileReader(new File(cmdHistoryFile)));
-			String line;
-			while((line = br.readLine()) != null){
-				if(line.startsWith(MARKER_FOR_HISTORY_CMD)){
-					cmdHistory.add(line.replaceFirst(MARKER_FOR_HISTORY_CMD, ""));
-				}
-			}
-			br.close();
-		} catch(IOException e){
-			if(debug > 0){
-				e.printStackTrace();
-			}
-		}
-		return cmdHistory;
-	}	
-
-	/** Merge set of opened files in trackSet with the files found in the history file.
-	 * */
-	private static void addHistoryFiles(TrackSet trackSet, String cmdHistoryFile, String MARKER_FOR_HISTORY_FILE, int debug){
-		LinkedHashSet<String> union= initRecentlyOpenedFiles(cmdHistoryFile, MARKER_FOR_HISTORY_FILE, debug);
-		LinkedHashSet<String> now= trackSet.getOpenedFiles();
-		for(String file : now){ // If a file is in the current track set and in the history file, put it last. I.e. last opened. 
-			if(union.contains(file)){
-				union.remove(file);
-			}
-			union.add(file);
-		}
-		trackSet.setOpenedFiles(union);
-	}
-	
-	/**Read the asciigenome history file to extract the list of opened files 
-	 * */
-	private static LinkedHashSet<String> initRecentlyOpenedFiles(String cmdHistoryFile, 
-			String MARKER_FOR_HISTORY_FILE, int debug){
-		LinkedHashSet<String> opened= new LinkedHashSet<String>();
-		try{
-			BufferedReader br= new BufferedReader(new FileReader(new File(cmdHistoryFile)));
-			String line;
-			while((line = br.readLine()) != null){
-				if(line.startsWith(MARKER_FOR_HISTORY_FILE)){
-					line= line.substring(MARKER_FOR_HISTORY_FILE.length(), line.length()).trim();
-					opened.add(line);
-				}
-			}
-			br.close();
-		} catch(IOException e){
-			if(debug > 0){
-				e.printStackTrace();
-			}
-		}
-		return opened;
-	}	
-	
 	/** Set some sensible defaults for track heights 
 	 * @throws SQLException 
 	 * @throws InvalidRecordException 
@@ -472,56 +378,80 @@ public class Main {
 		}
 	}
 	
-	/**Write the history of commands to given file.
-	 * Note that the existing history file is completely overwritten.
+	/** On shutdown, prepare and write the history file. Not that the existing 
+	 * yaml file is overwritten.
+	 * @param session 
 	 * */
-	private static void writeHistory(final History cmdHistory, final Set<String> fileHistory, final String cmdHistoryFile, final String MARKER_FOR_HISTORY_FILE, 
-				final String MARKER_FOR_HISTORY_CMD, final GenomicCoordsHistory gch){
-		
+	private static void writeYamlHistory(ASCIIGenomeHistory current, 
+										 final History cmdHistory, 
+										 TrackSet trackSet,
+			                             final GenomicCoordsHistory gch 
+			                             ){
+	
 		Runtime.getRuntime().addShutdownHook(new Thread(new Runnable() {
 			public void run() {
-	            
+
+			ASCIIGenomeHistory newYamlHistory= null;
+			try {
+				newYamlHistory = new ASCIIGenomeHistory(null);
+			} catch (IOException e1) {
+				e1.printStackTrace();
+			} 
+			try{
 				// List of commands
 				ListIterator<Entry> iter = cmdHistory.entries();
-				List<String>lastHist= new ArrayList<String>();
+				List<String>lastCommands= new ArrayList<String>();
 				int max_cmds= 2000; // Maximum number of commands to write out to asciigenomo_history.
 				while(iter.hasNext()){
 					if(max_cmds == 0){
 						break;
 					}
 					max_cmds--;
-					lastHist.add(iter.next().value().toString());
+					lastCommands.add(iter.next().value().toString());
 				}
+				newYamlHistory.setCommands(lastCommands);
 				
 				// List of files
-				List<String>lastFiles= new ArrayList<String>(fileHistory);
+				List<String>lastFiles= new ArrayList<String>(trackSet.getOpenedFiles());
 				int max_files= 200; // Maximum number of files to write out to asciigenomo_history.
 				lastFiles= lastFiles.subList(Math.max(0, lastFiles.size() - max_files), lastFiles.size());
+				newYamlHistory.setFiles(lastFiles);
+				
+				// Positions
+				List<String> lastPos= gch.prepareHistoryForHistoryFile(newYamlHistory.getFileName(), 100);
+				newYamlHistory.setPositions(lastPos);
 
-				try{
-					
-					StringBuilder sb= new StringBuilder();
-					
-					for(String cmd : lastHist){
-						sb.append(MARKER_FOR_HISTORY_CMD + cmd + "\n");
-					}
-					for(String file : lastFiles){
-						sb.append(MARKER_FOR_HISTORY_FILE + file + "\n");
-					}
-					for(String pos : gch.prepareHistoryForHistoryFile(new File(cmdHistoryFile), 100)){
-						sb.append(pos + "\n"); // Positions
-					}
-
-					BufferedWriter wr= new BufferedWriter(new FileWriter(new File(cmdHistoryFile)));
-					wr.write(sb.toString());
-					wr.close();
-				} catch (IOException e) {
-					System.err.println("Unable to write history to " + cmdHistoryFile);
+//				// Sessions
+//				Map<Integer, List<String>> currentSessions = current.getSessions();
+//				if(currentSessions == null){
+//					currentSessions= new LinkedHashMap<Integer, List<String>>(); 
+//				}
+//				boolean found= false;
+//				System.err.println(currentSessions);
+//				for(int x : currentSessions.keySet()){
+//					if(currentSessions.get(x).equals(trackSet.getFilenameList())){
+//						found= true;
+//						break;
+//					}
+//				}
+//				System.err.println(trackSet.getFilenameList());
+//				if( ! found && ! trackSet.getFilenameList().isEmpty()){
+//					int nextKey= currentSessions.keySet().size() + 1;
+//					currentSessions.put(nextKey, trackSet.getFilenameList());
+//				}
+//				newYamlHistory.setSessions(currentSessions);
+				
+				// Write yaml
+				newYamlHistory.write();
+				
+			} catch (Exception e) {
+				e.printStackTrace();
+					System.err.println("Unable to write history to " + newYamlHistory.getFileName());
 				}
-	        }
+			}
 	    }, "Shutdown-thread"));
 	}
-		
+	
 	/** On exit print a message informing a new version of ASCIIGenome is available
 	 * */
 	private static void messageVersion(final boolean noFormat) throws IOException, InvalidColourException{